requirements.txt
django
djangorestframework
langchain_openai
pinecone
langchain_pinecone
langchain_community